refactor: 게임모드 추상화 제거 후 정적팩토리 메서드로 수정
